+++ lib/dy_index/focus/index.dart
@@ -1,184 +1,15 @@
-/**
- * @discripe: 关注
- */
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:fluttertoast/fluttertoast.dart';
 
-import 'develop.dart';
+import '../../base.dart';
 
-class FocusPage extends StatelessWidget {
+class FocusPage extends StatefulWidget {
   @override
-  Widget build(BuildContext context) {
-    return CustomScrollView(
-        // physics: BouncingScrollPhysics(),
-        slivers: <Widget>[
-          SliverPersistentHeader(
-            pinned: true,
-            delegate: SliverCustomHeaderDelegate(
-              title: '关注（现为新加功能测试页面）',
-              collapsedHeight: 40,
-              expandedHeight: 300,
-              paddingTop: MediaQuery.of(context).padding.top,
-              coverImgUrl: 'http://r.photo.store.qq.com/psb?/V14dALyK4PrHuj/iEyZ0zEnZQmbqfs.BUMe9Visc92Tqohh0OKKP0JOtVU!/r/dMMAAAAAAAAA'
-            ),
-          ),
-          SliverToBoxAdapter(
-            child: FilmContent(),
-          )
-        ],
-      );
-  }
+  _FocusPage createState() => _FocusPage();
 }
 
-/// [SliverPersistentHeaderDelegate]在原有Widget的基础上封装了滚动坐标，在build方法中的shrinkOffset参数
-/// 这里利用继承原生组件实现了随滚动改变header透明度的效果
-class SliverCustomHeaderDelegate extends SliverPersistentHeaderDelegate {
-  final double collapsedHeight;
-  final double expandedHeight;
-  final double paddingTop;
-  final String coverImgUrl;
-  final String title;
-  String statusBarMode = 'dark';
-
-  SliverCustomHeaderDelegate({
-    this.collapsedHeight,
-    this.expandedHeight,
-    this.paddingTop,
-    this.coverImgUrl,
-    this.title,
-  });
-
-  @override
-  double get minExtent => this.collapsedHeight + this.paddingTop;
-
-  @override
-  double get maxExtent => this.expandedHeight;
-
-  @override
-  bool shouldRebuild(SliverPersistentHeaderDelegate oldDelegate) {
-    return true;
-  }
-
-  void updateStatusBarBrightness(shrinkOffset) {
-    if(shrinkOffset <= 50 && this.statusBarMode == 'dark') {
-      this.statusBarMode = 'light';
-      SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-        statusBarBrightness: Brightness.light,
-        statusBarIconBrightness: Brightness.light,
-      ));
-    } else if(shrinkOffset > 50 && this.statusBarMode == 'light') {
-      this.statusBarMode = 'dark';
-      SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-        statusBarBrightness: Brightness.dark,
-        statusBarIconBrightness: Brightness.dark,
-      ));
+class _FocusPage extends State<FocusPage> with DYBase {
+   @override
+    Widget build(BuildContext context) {
+      return Scaffold();
     }
-  }
-
-  Color makeStickyHeaderBgColor(shrinkOffset) {
-    final int alpha = (shrinkOffset / (this.maxExtent - this.minExtent) * 255).clamp(0, 255).toInt();
-    return Color.fromARGB(alpha, 255, 255, 255);
-  }
-
-  Color makeStickyHeaderTextColor(shrinkOffset, isIcon) {
-    if(shrinkOffset <= 50) {
-      return isIcon ? Colors.white : Colors.transparent;
-    } else {
-      final int alpha = (shrinkOffset / (this.maxExtent - this.minExtent) * 255).clamp(0, 255).toInt();
-      return Color.fromARGB(alpha, 0, 0, 0);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context, double shrinkOffset, bool overlapsContent) {
-    this.updateStatusBarBrightness(shrinkOffset);
-    return Container(
-      height: this.maxExtent,
-      width: MediaQuery.of(context).size.width,
-      child: Stack(
-        fit: StackFit.expand,
-        children: <Widget>[
-          Container(child: Image.network(this.coverImgUrl, fit: BoxFit.cover)),
-          Positioned(
-            left: 0,
-            top: this.maxExtent / 2,
-            right: 0,
-            bottom: 0,
-            child: Container(
-              decoration: BoxDecoration(
-                gradient: LinearGradient(
-                  begin: Alignment.topCenter,
-                  end: Alignment.bottomCenter,
-                  colors: [
-                    Color(0x00000000),
-                    Color(0x90000000),
-                  ],
-                ),
-              ),
-            ),
-          ),
-          Positioned(
-            left: 0,
-            right: 0,
-            top: 0,
-            child: Container(
-              color: this.makeStickyHeaderBgColor(shrinkOffset),
-              child: SafeArea(
-                bottom: false,
-                child: Container(
-                  height: this.collapsedHeight,
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: <Widget>[
-                      IconButton(
-                        icon: Icon(
-                          Icons.arrow_back_ios,
-                          color: this.makeStickyHeaderTextColor(shrinkOffset, true),
-                        ),
-                        onPressed: () => Fluttertoast.showToast(msg: '没有上一页了'),
-                      ),
-                      Text(
-                        this.title,
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.w500,
-                          color: this.makeStickyHeaderTextColor(shrinkOffset, false),
-                        ),
-                      ),
-                      IconButton(
-                        icon: Icon(
-                          Icons.share,
-                          color: this.makeStickyHeaderTextColor(shrinkOffset, true),
-                        ),
-                        onPressed: () {},
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class FilmContent extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.all(16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          DevelopTest(),
-          SizedBox(
-            height: 700,
-          )
-        ],
-      ),
-    );
-  }
 }
